fix(doodad): cast size_t to int in getMaxVariation() method for type consistency

+++ source/doodad_brush.h
@@ -101,7 +101,7 @@ public:
 		return look_id;
 	}
 	virtual int getMaxVariation() const {
-		return alternatives.size();
+		return static_cast<int>(alternatives.size());
 	}
 	virtual std::string getName() const {
 		return name;
